lib/__init__.py
import torch.nn as nn
import torch.optim as optim
import torch.utils.data as torch_data
import torchvision.transforms as transforms
import pytorch_lightning as pl
import pytorch_lightning.loggers as pl_loggers
import pytorch_lightning.callbacks as pl_callbacks

from nip import nip

from . import datasets
from . import lightning_modules
from . import losses
from . import models

nip(nn)
nip(optim)
nip(torch_data)
nip(transforms)
nip(pl)
nip(pl_loggers)
nip(pl_callbacks)
