requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.0
torchvision==0.14.0
pytorch-lightning==1.8.3
neptune-client==0.16.13
compressai==1.2.1
vector-quantize-pytorch==0.10.14
opencv-python==4.5.5.64
nip-config==0.8.3
